+++ FakeServer/Authentication/JWT/TokenLogoutMiddleware.cs
@@ -2,13 +2,32 @@
 using Microsoft.Extensions.Options;
 using System;
 using System.Collections.Generic;
+using System.IdentityModel.Tokens.Jwt;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace FakeServer.Authentication.Jwt
 {
     public class TokenBlacklistService
     {
-        public List<string> Headers { get; } = new List<string>();
+        private readonly List<string> _headers = new List<string>();
+
+        public void BlacklistHeader(string header)
+        {
+            if (GetJtiFromToken(header, out var jti))
+                _headers.Add(jti);
+        }
+
+        public bool IsBlacklisted(string header) => GetJtiFromToken(header, out var jti) ? _headers.Contains(jti) : false;
+
+        private bool GetJtiFromToken(string header, out string jti)
+        {
+            header = header.Replace("Bearer ", "");
+
+            var jsonToken = new JwtSecurityTokenHandler().ReadToken(header) as JwtSecurityToken;
+            jti = jsonToken?.Claims.FirstOrDefault(claim => claim.Type == "jti")?.Value;
+            return !string.IsNullOrEmpty(jti);
+        }
     }
 
     public class TokenLogoutMiddleware
@@ -41,7 +60,7 @@ namespace FakeServer.Authentication.Jwt
             var blacklistService = context.RequestServices.GetService(typeof(TokenBlacklistService)) as TokenBlacklistService;
 
             var header = context.Request.Headers["Authorization"];
-            blacklistService.Headers.Add(header.ToString());
+            blacklistService.BlacklistHeader(header.ToString());
 
             return Task.FromResult(0);
         }