+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ### [Unreleased]
-*
+* JWT blacklist to check token identifier from JWT ID (jti)
 
 ### [0.6.0] - 2017-11-30
 * Caching of unchanged resources with ETag and If-None-Match headers